// microservices/warehouse/src/shared/domain/enums/order-state.enum.ts
export enum OrderState {
  CREATED = 'CREATED',
  RESERVING = 'RESERVING',
  RESERVED = 'RESERVED',
  VALIDATING = 'VALIDATING',
  REPLENISHING = 'REPLENISHING',
  CANCELING = 'CANCELING',
  CANCELED = 'CANCELED',
  DISPATCHING = 'DISPATCHING',
  DISPATCHED = 'DISPATCHED',
  DELIVERED = 'DELIVERED',
}